Hotfix 1:  - Fix handling of integer literals in ACSL annotations (Integers in ACSL annotations are unbouded mathematical integers)  - Fix IndexOutOfBoundsException in case ACSL assertion is placed after return statement  - Make note of possible NullPointerException (not sure how to handle right now, should not happen often)
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/acsl-bachelor@36264 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/core/algorithm/acsl/ACSLTermToCExpressionVisitor.java
@@ -8,7 +8,6 @@
 
 package org.sosy_lab.cpachecker.core.algorithm.acsl;
 
-import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.Map;
 import org.sosy_lab.common.log.LogManager;
@@ -191,7 +190,7 @@ public class ACSLTermToCExpressionVisitor {
           new CIntegerLiteralExpression(
               FileLocation.DUMMY,
               CNumericTypes.INT,
-              BigInteger.valueOf(integerLiteral.getLiteral()));
+              integerLiteral.getLiteral());
       cache.put(integerLiteral, result);
     }
     return result;
